refactor timeline and actionbar business logic to component level
