+++ src/component/actionbar/index.tsx
@@ -9,32 +9,46 @@ import ComposeEditor from 'component/editor/compose';
 import { EditorShape } from 'component/editor/compose';
 import { INewUpdate } from 'module/interface/update';
 import { useMutation, useQueryClient } from 'react-query';
-import { initialValueEmpty } from 'component/editor/config/initialValues';
-import { serialize } from 'module/serialize';
 import * as api from 'module/api';
 
+import { initialValueEmpty } from 'component/editor/config/initialValues';
+import { Search } from '@venturemark/numnum';
+import { serialize } from 'module/serialize';
+import { get } from 'module/store';
+import { useEditor } from 'component/editor/compose';
+
 interface ActionBarProps extends DefaultActionBarProps {
-  audienceId: string;
   organizationId: string;
   timelineId: string;
   userId: string;
-  errorMessage: string;
-  progress: number;
-  editorShape: EditorShape;
-  setEditorShape: React.Dispatch<React.SetStateAction<EditorShape>>;
 }
 
 function ActionBar(props: ActionBarProps) {
   const {
-    audienceId,
     organizationId,
     timelineId,
     userId,
-    errorMessage,
-    progress,
-    editorShape,
-    setEditorShape,
   } = props;
+
+  const store = get('composeEditor.content') ?? '';
+  const initialValue =
+    store !== '' ? JSON.parse(store) : initialValueEmpty;
+  const hasContentDefault =
+    serialize(initialValue) === '' ||
+    serialize(initialValue) === undefined
+      ? undefined
+      : 'hasContent';
+  const defaultNumber = Search(serialize(initialValue)) ?? 0;
+  const defaultProgress = serialize(initialValue).length;
+
+  const { editorShape, setEditorShape } = useEditor({
+    value: initialValue,
+    hasContent: hasContentDefault,
+    numberValue: defaultNumber[0],
+    progress: defaultProgress,
+  });
+
+  const audienceId = '1'
 
   const queryClient = useQueryClient();
 
@@ -52,7 +66,7 @@ function ActionBar(props: ActionBarProps) {
 
   const handleAddUpdate = () => {
     if (!timelineId) {
-      const error = 'Please create a timeline';
+      const error = 'Please select a timeline';
       setEditorShape({ ...editorShape, error });
       return;
     }
@@ -86,7 +100,7 @@ function ActionBar(props: ActionBarProps) {
       JSON.stringify(initialValueEmpty),
     );
     //reset editor
-    const resetEditor = {
+    const resetEditor:EditorShape = {
       value: initialValueEmpty,
       string: '',
       hasContent: undefined,
@@ -113,13 +127,13 @@ function ActionBar(props: ActionBarProps) {
       sendUpdate={{
         handleClick: () => handleAddUpdate(),
       }}
-      error={errorMessage ? 'hasError' : undefined}
-      text={normalize(progress) > 0 ? 'hasText' : undefined}
+      error={editorShape.error ? 'hasError' : undefined}
+      text={normalize(editorShape.progress) > 0 ? 'hasText' : undefined}
       timelineSelected={timelineSelected}
       timelineSelect={{
         onClick: () => setTimelineSelected(true),
       }}
-      errorMessage={errorMessage}
+      errorMessage={editorShape.error}
       isActive={isActive}
       textContainer={{
         render: () => (
